requirements.txt
groq==0.9.0
openai==1.40.2
pytesseract==0.3.10
pillow==10.4.0